wip: migrate dump
