+++ cmd/dump/dump.go
@@ -6,9 +6,7 @@ import (
 
 	root "github.com/koh-sh/codebuild-multirunner/cmd"
 	mr "github.com/koh-sh/codebuild-multirunner/internal/multirunner"
-	"github.com/koh-sh/codebuild-multirunner/internal/types"
 	"github.com/spf13/cobra"
-	"gopkg.in/yaml.v3"
 )
 
 // dumpCmd represents the dump command
@@ -20,7 +18,7 @@ var dumpCmd = &cobra.Command{
 		if err != nil {
 			log.Fatal(err)
 		}
-		conf, err := dumpConfig(bc)
+		conf, err := mr.DumpConfig(bc)
 		if err != nil {
 			log.Fatal(err)
 		}
@@ -30,13 +28,4 @@ var dumpCmd = &cobra.Command{
 
 func init() {
 	root.RootCmd.AddCommand(dumpCmd)
-}
-
-// dump read config with environment variables inserted
-func dumpConfig(bc types.BuildConfig) (string, error) {
-	d, err := yaml.Marshal(&bc)
-	if err != nil {
-		return "", err
-	}
-	return string(d), nil
 }
